--- a/cmd/forwarder/fixer.go
+++ b/cmd/forwarder/fixer.go
@@ -6,6 +6,7 @@ import (
 	"io"
 	"net/http"
 	"strconv"
+	"sync"
 
 	"github.com/bmizerany/lpx"
 )
@@ -19,46 +20,63 @@ const (
 var nilVal = []byte(`- `)
 var queryParams = []string{"index", "source", "sourcetype"}
 
+var pool = sync.Pool{
+	New: func() interface{} {
+		return &bytes.Buffer{}
+	},
+}
+
+func getBuffer() *bytes.Buffer {
+	return pool.Get().(*bytes.Buffer)
+}
+
+func putBuffer(b *bytes.Buffer) {
+	b.Reset() // ensure that the buffer is ready for re-use
+	pool.Put(b)
+}
+
 // Get metadata from the http request.
 // Returns an empty byte array if there isn't any.
-func getMetadata(req *http.Request, cred *credential, metadataId string) ([]byte, bool) {
-	var metadataWriter bytes.Buffer
-	var foundMetadata bool
+func getMetadata(req *http.Request, cred *credential, metadataId string) []byte {
+	if metadataId == "" { // short circuit when metadataId is empty
+		return nil
+	}
+
+	metadataWriter := getBuffer()
+	defer putBuffer(metadataWriter)
+	writeSDElementStart := func() {
+		metadataWriter.WriteString("[")
+		metadataWriter.WriteString(metadataId)
+	}
+
 	// Calculate metadata query parameters
-	if metadataId != "" {
-		for _, k := range queryParams {
-			v := req.FormValue(k)
-			if v != "" {
-				if !foundMetadata {
-					metadataWriter.WriteString("[")
-					metadataWriter.WriteString(metadataId)
-					foundMetadata = true
-				}
-				metadataWriter.WriteString(" ")
-				metadataWriter.WriteString(k)
-				metadataWriter.WriteString("=\"")
-				metadataWriter.WriteString(v)
-				metadataWriter.WriteString("\"")
+	for _, k := range queryParams {
+		if v := req.FormValue(k); v != "" {
+			if metadataWriter.Len() == 0 {
+				writeSDElementStart()
 			}
-		}
-
-		// Add metadata about the credential if it is deprecated
-		if cred != nil && cred.Deprecated == true {
-			if !foundMetadata {
-				metadataWriter.WriteString("[")
-				metadataWriter.WriteString(metadataId)
-				foundMetadata = true
-			}
-			metadataWriter.WriteString(" fields=\"{'credential_deprecated': true, 'credential_name': '")
-			metadataWriter.WriteString(cred.Name)
-			metadataWriter.WriteString("'}\"")
-		}
-
-		if foundMetadata {
-			metadataWriter.WriteString("]")
+			metadataWriter.WriteString(" ")
+			metadataWriter.WriteString(k)
+			metadataWriter.WriteString("=\"")
+			metadataWriter.WriteString(v)
+			metadataWriter.WriteString("\"")
 		}
 	}
-	return metadataWriter.Bytes(), foundMetadata
+
+	// Add metadata about the credential if it is deprecated
+	if cred != nil && cred.Deprecated == true {
+		if metadataWriter.Len() == 0 {
+			writeSDElementStart()
+		}
+		metadataWriter.WriteString(` fields="credential_deprecated=true,credential_name=`)
+		metadataWriter.WriteString(cred.Name)
+		metadataWriter.WriteString(`"`)
+	}
+
+	if metadataWriter.Len() > 0 {
+		metadataWriter.WriteString("]")
+	}
+	return metadataWriter.Bytes()
 }
 
 // Fix function to convert post data to length prefixed syslog frames
@@ -68,15 +86,18 @@ func getMetadata(req *http.Request, cred *credential, metadataId string) ([]byte
 // * byte array of syslog data.
 // * error if something went wrong.
 func fix(req *http.Request, r io.Reader, remoteAddr string, logplexDrainToken string, metadataId string, cred *credential) (bool, int64, []byte, error) {
-	var messageWriter bytes.Buffer
-	var messageLenWriter bytes.Buffer
+	messageWriter := getBuffer()
+	messageLenWriter := getBuffer()
+	defer func() {
+		putBuffer(messageWriter)
+		putBuffer(messageLenWriter)
+	}()
 
-	metadataBytes, hasMetadata := getMetadata(req, cred, metadataId)
+	metadataBytes := getMetadata(req, cred, metadataId)
 
 	lp := lpx.NewReader(bufio.NewReader(r))
-	numLogs := int64(0)
-	for lp.Next() {
-		numLogs++
+	var numLogs int64
+	for ; lp.Next(); numLogs++ {
 		header := lp.Header()
 
 		// LEN SP PRI VERSION SP TIMESTAMP SP HOSTNAME SP APP-NAME SP PROCID SP MSGID SP STRUCTURED-DATA MSG
@@ -100,12 +121,9 @@ func fix(req *http.Request, r io.Reader, remoteAddr string, logplexDrainToken st
 		messageWriter.WriteString("\"]")
 
 		// Write metadata
-		if hasMetadata {
-			messageWriter.Write(metadataBytes)
-		}
+		messageWriter.Write(metadataBytes)
 
-		b := lp.Bytes()
-		if len(b) >= 2 && bytes.Equal(b[0:2], nilVal) {
+		if b := lp.Bytes(); len(b) >= 2 && bytes.Equal(b[0:2], nilVal) {
 			messageWriter.Write(b[1:])
 		} else if len(b) > 0 {
 			if b[0] != '[' {
@@ -116,8 +134,8 @@ func fix(req *http.Request, r io.Reader, remoteAddr string, logplexDrainToken st
 
 		messageLenWriter.WriteString(strconv.Itoa(messageWriter.Len()))
 		messageLenWriter.WriteString(" ")
-		messageWriter.WriteTo(&messageLenWriter)
+		messageWriter.WriteTo(messageLenWriter)
 	}
 
-	return hasMetadata, numLogs, messageLenWriter.Bytes(), lp.Err()
+	return len(metadataBytes) > 0, numLogs, messageLenWriter.Bytes(), lp.Err()
 }