This uses a sync.Pool to keep a series of already initialized
bytes.Buffers and removed some un-needed checks.

Here are the changes courtesy of benchcmp:

|benchmark|old ns/op|new ns/op|delta|
|--|--|--|--|
|BenchmarkFixNoSD-8|3516|3161|-10.10%|
|BenchmarkFixSD-8|2472|2277|-7.89%|
|BenchmarkFixSDWithMetadata-8|2695|2525|-6.31%|

|benchmark|old MB/s|new MB/s|speedup|
|--|--|--|--|
|BenchmarkFixNoSD-8|38.95|43.33|1.11x|
|BenchmarkFixSD-8|44.49|48.29|1.09x|
|BenchmarkFixSDWithMetadata-8|40.82|43.56|1.07x|

|benchmark|old allocs|new allocs|delta|
|--|--|--|--|
|BenchmarkFixNoSD-8|44|39|-11.36%|
|BenchmarkFixSD-8|29|24|-17.24%|
|BenchmarkFixSDWithMetadata-8|32|27|-15.62%|

|benchmark|old bytes|new bytes|delta|
|--|--|--|--|
|BenchmarkFixNoSD-8|5992|5449|-9.06%|
|BenchmarkFixSD-8|5760|5185|-9.98%|
|BenchmarkFixSDWithMetadata-8|5872|5298|-9.78%|