@freeformz 2019-04-04
@sheax0r When this is eventually deployed, I suggest rolling it out to a single indexer and let it run for a bit. I suggest that approach for deploying any change though.